refactor: 변수명 수정, SurveyValidator, SurveyQuestionService 빈 등록 제거로 생성자 사용
