+++ src/main/java/com/example/surveyapp/domain/survey/application/OptionsService.java
@@ -1,13 +1,14 @@
 package com.example.surveyapp.domain.survey.application;
 
-import com.example.surveyapp.domain.survey.controller.dto.request.OptionCreateRequestDto;
-import com.example.surveyapp.domain.survey.controller.dto.request.OptionUpdateRequestDto;
-import com.example.surveyapp.domain.survey.controller.dto.response.OptionResponseDto;
+import com.example.surveyapp.domain.survey.presentation.dto.request.OptionCreateRequestDto;
+import com.example.surveyapp.domain.survey.presentation.dto.request.OptionUpdateRequestDto;
+import com.example.surveyapp.domain.survey.presentation.dto.response.OptionResponseDto;
 import com.example.surveyapp.domain.survey.domain.SurveyValidator;
 import com.example.surveyapp.domain.survey.domain.model.entity.Options;
 import com.example.surveyapp.domain.survey.domain.model.entity.Question;
 import com.example.surveyapp.domain.survey.domain.model.entity.Survey;
 import com.example.surveyapp.domain.survey.domain.service.SurveyQuestionService;
+import com.example.surveyapp.domain.user.domain.model.UserRoleEnum;
 import com.example.surveyapp.global.reader.UserReader;
 
 import lombok.RequiredArgsConstructor;
@@ -22,15 +23,15 @@ import java.util.stream.Collectors;
 public class OptionsService {
 
     private final UserReader userReader;
-    private final SurveyValidator surveyValidator;
-    private final SurveyQueryService surveyQuestionQueryService;
-    private final SurveyQuestionService surveyQuestionService;
+    private final SurveyValidator surveyValidator = new SurveyValidator();
+    private final SurveyQueryService surveyQueryService;
+    private final SurveyQuestionService surveyQuestionService = new SurveyQuestionService();
 
     @Transactional
     public OptionResponseDto createOption(Long userId, Long surveyId, Long questionId, OptionCreateRequestDto requestDto){
 
         userReader.validateUserIdOrThrow(userId);
-        Survey survey = surveyQuestionQueryService.findSurvey(surveyId);
+        Survey survey = surveyQueryService.findSurvey(surveyId);
         Question question = surveyQuestionService.getQuestionById(survey, questionId);
 
         surveyValidator.validateUpdatable(userId, survey);
@@ -46,15 +47,15 @@ public class OptionsService {
     }
 
     //(inprogress일때는 모두, 다른 상태에는 관리자랑 해당 설문 출제자만 조회가능)
-    //in progress이고 유저가 참여자권한인 경우 이미 참여했는지 확인해야함. - 응답 도메인 생성 후 수정
+    @Transactional(readOnly = true)
     public List<OptionResponseDto> getOptions(Long userId, Long surveyId, Long questionId){
 
         userReader.validateUserIdOrThrow(userId);
 
-        Survey survey = surveyQuestionQueryService.findSurvey(surveyId);
+        Survey survey = surveyQueryService.findSurvey(surveyId);
         Question question = surveyQuestionService.getQuestionById(survey, questionId);
 
-        surveyValidator.validateQuestionAccess(userId, survey);
+        surveyValidator.validateQuestionAccess(userId, survey, userReader.validateUserRole(userId, UserRoleEnum.SURVEYEE));
 
         List<Options> optionsList = question.getOptions();
 
@@ -72,7 +73,7 @@ public class OptionsService {
 
         userReader.validateUserIdOrThrow(userId);
 
-        Survey survey = surveyQuestionQueryService.findSurvey(surveyId);
+        Survey survey = surveyQueryService.findSurvey(surveyId);
         Question question = surveyQuestionService.getQuestionById(survey, questionId);
         surveyValidator.validateUpdatable(userId, survey);
 
@@ -89,7 +90,7 @@ public class OptionsService {
 
         userReader.validateUserIdOrThrow(userId);
 
-        Survey survey = surveyQuestionQueryService.findSurvey(surveyId);
+        Survey survey = surveyQueryService.findSurvey(surveyId);
         Question question = surveyQuestionService.getQuestionById(survey, questionId);
 
         surveyValidator.validateDeletable(userId, survey);
